refactor: Pruebas con layout para cada año del historial
